[#17]기존 member에 있던 내용들 members로 옮겼습니다
