--- src/main/java/clonecoding/tinder/config/S3config.java
@@ -1,33 +1,0 @@
-//package clonecoding.tinder.config;
-//
-//import com.amazonaws.auth.AWSCredentials;
-//import com.amazonaws.auth.AWSStaticCredentialsProvider;
-//import com.amazonaws.auth.BasicAWSCredentials;
-//import com.amazonaws.services.s3.AmazonS3;
-//import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-//import org.springframework.beans.factory.annotation.Value;
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.Configuration;
-//
-//@Configuration
-//public class S3config {
-//    @Value("${cloud.aws.credentials.accessKey}")
-//    private String accessKey;
-//
-//    @Value("${cloud.aws.credentials.secretKey}")
-//    private String secretKey;
-//
-//    @Value("${cloud.aws.region.static}")
-//    private String region;
-//
-//    @Bean
-//    public AmazonS3 amazonS3Client() {
-//        AWSCredentials credentials = new BasicAWSCredentials(accessKey, secretKey);
-//
-//        return AmazonS3ClientBuilder
-//                .standard()
-//                .withCredentials(new AWSStaticCredentialsProvider(credentials))
-//                .withRegion(region)
-//                .build();
-//    }
-//}
