+++ src/main/java/clonecoding/tinder/like/entity/Likes.java
@@ -1,6 +1,5 @@
 package clonecoding.tinder.like.entity;
 
-import clonecoding.tinder.member.entity.Member;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
