+++ src/main/java/clonecoding/tinder/like/service/LikeService.java
@@ -4,8 +4,8 @@ import clonecoding.tinder.like.dto.LikeResponseDto;
 import clonecoding.tinder.like.dto.LikedMemberResponseDto;
 import clonecoding.tinder.like.entity.Likes;
 import clonecoding.tinder.like.repository.LikeRepository;
-import clonecoding.tinder.member.entity.Member;
-import clonecoding.tinder.member.repository.MemberRepository;
+import clonecoding.tinder.members.entity.Member;
+import clonecoding.tinder.members.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
