+++ src/main/java/clonecoding/tinder/members/dto/MemberSignupRequestDto.java
@@ -1,4 +1,4 @@
-package clonecoding.tinder.member.dto;
+package clonecoding.tinder.members.dto;
 
 import lombok.Getter;
