+++ src/main/java/clonecoding/tinder/members/entity/MemberRoleEnum.java
@@ -1,4 +1,4 @@
-package clonecoding.tinder.member.entity;
+package clonecoding.tinder.members.entity;
 
 public enum MemberRoleEnum {
